scrap CreateNewTool (#128)
* removed createnewtool and related code

* fixed path registry ckpt

* fixed unit tests and tidy up
