+++ mdagent/__init__.py
@@ -1,3 +1,3 @@
-from .mainagent import Evaluator, MDAgent
+from .agent import Evaluator, MDAgent
 
 __all__ = ["MDAgent", "Evaluator"]
