--- mdagent/subagents/__init__.py
@@ -1,14 +1,0 @@
-from .agents import Action, Critic, Curriculum, MemoryManager, SkillManager
-from .subagent_fxns import Iterator
-from .subagent_setup import SubAgentInitializer, SubAgentSettings
-
-__all__ = [
-    "Action",
-    "Critic",
-    "Curriculum",
-    "SkillManager",
-    "Iterator",
-    "SubAgentInitializer",
-    "SubAgentSettings",
-    "MemoryManager",
-]
